Fixed desync between status bar and stored level. Fixed duplicated level info in level storage.
Prior to the desync fix the current exp of the status bar would always be 1 less than the current exp being saved. Although menial a fix was required for the extension to work as intended. The extension would also duplicate the level information on the first save leading to something like1,22,100,1,22,100. While this was menial as well it would cause issues with displaying level information if left unresolved.